resolved linking issues.
Removed redundant plugin components and added Infill as a direct dependency. Also, updated the plugin proxy to handle the invocation and broadcasting tasks previously handled by the deleted plugin components. Updated and moved function 'prep_client_context' to be used within plugin proxy.

Added a direct #include for "infill.h" in "broadcasts.h" and updated corresponding cmake to reflect the changes.

Commented out the inclusion of "plugins/slots.h" in "LayerPlan.cpp" to resolve linking issues.

These changes greatly simplify the code structure, improve clarity and ensure successful linking during project build.

Contributes to CURA-10916
