+++ include/plugins/broadcasts.h
@@ -9,6 +9,7 @@
 #include "plugins/converters.h"
 
 #include <agrpc/asio_grpc.hpp>
+#include <range/v3/utility/semiregular_box.hpp>
 
 #include <type_traits>
 
@@ -24,10 +25,9 @@ struct is_broadcast_channel
 template<plugins::v0::SlotID T1, plugins::v0::SlotID T2>
 inline constexpr bool is_broadcast_channel_v = is_broadcast_channel<T1, T2>::value;
 
-template<class T = void, class C = empty>
+template<class T, class C = empty>
 struct broadcast_stub
 {
-    using stub_type = slots::broadcast::v0::BroadcastService::Stub;
     using derived_type = T;
     friend derived_type;
 
@@ -40,12 +40,11 @@ private:
     C request_{};
 };
 
-template<v0::SlotID S>
+template<v0::SlotID S, class Stub>
 requires is_broadcast_channel_v<S, v0::SlotID::SETTINGS_BROADCAST>
-struct broadcast_rpc : public broadcast_stub<broadcast_rpc<S>, broadcast_settings_request>
+struct broadcast_rpc : public broadcast_stub<broadcast_rpc<S, Stub>, broadcast_settings_request>
 {
-    using base_type = broadcast_stub<broadcast_rpc<S>>;
-    using ClientRPC = agrpc::ClientRPC<&base_type::stub_type::PrepareAsyncBroadcastSettings>;
+    using ClientRPC = agrpc::ClientRPC<&Stub::PrepareAsyncBroadcastSettings>;
 };
