+++ include/plugins/pluginproxy.h
@@ -5,9 +5,6 @@
 #define PLUGINS_PLUGINPROXY_H
 
 #include "Application.h"
-#include "components/broadcast.h"
-#include "components/common.h"
-#include "components/invoke.h"
 #include "cura/plugins/slots/broadcast/v0/broadcast.grpc.pb.h"
 #include "cura/plugins/slots/broadcast/v0/broadcast.pb.h"
 #include "cura/plugins/slots/handshake/v0/handshake.grpc.pb.h"
@@ -34,6 +31,9 @@
 #include <spdlog/spdlog.h>
 
 #include <chrono>
+#include <coroutine>
+#include <memory>
+#include <optional>
 #include <string>
 #include <thread>
 
@@ -56,17 +56,18 @@ namespace cura::plugins
 template<plugins::v0::SlotID SlotID, utils::CharRangeLiteral SlotVersionRng, class Stub, class ValidatorTp, typename RequestTp, typename ResponseTp>
 class PluginProxy
 {
-public:
     // type aliases for easy use
     using value_type = typename ResponseTp::native_value_type;
     using validator_type = ValidatorTp;
-
     using req_converter_type = RequestTp;
     using rsp_converter_type = ResponseTp;
+    using rsp_msg_type = typename ResponseTp::value_type;
+    using invoke_stub_t = Stub;
+    using broadcast_stub_t = slots::broadcast::v0::BroadcastService::Stub;
 
-    using invoke_component_t = PluginProxyInvokeComponent<Stub, req_converter_type, rsp_converter_type>;
-    using broadcast_component_t = PluginProxyBroadcastComponent<SlotID>;
-
+    ranges::semiregular_box<invoke_stub_t> invoke_stub_; ///< The gRPC Invoke stub for communication.
+    ranges::semiregular_box<broadcast_stub_t> broadcast_stub_; ///< The gRPC Broadcast stub for communication.
+public:
     /**
      * @brief Constructs a PluginProxy object.
      *
@@ -81,8 +82,8 @@ public:
     constexpr PluginProxy() = default;
 
     explicit PluginProxy(std::shared_ptr<grpc::Channel> channel)
-        : invoke_component_{ slot_info_, plugin_info_, channel }
-        , broadcast_component_{ slot_info_, plugin_info_, channel }
+        : invoke_stub_{ channel }
+        , broadcast_stub_{ channel }
     {
         // Connect to the plugin and exchange a handshake
         agrpc::GrpcContext grpc_context;
@@ -96,21 +97,21 @@ public:
             {
                 using RPC = agrpc::ClientRPC<&slots::handshake::v0::HandshakeService::Stub::PrepareAsyncCall>;
                 grpc::ClientContext client_context{};
-                prep_client_context(client_context, *slot_info_);
+                prep_client_context(client_context, slot_info_);
 
                 // Construct request
                 handshake_request handshake_req;
-                handshake_request::value_type request{ handshake_req(*slot_info_) };
+                handshake_request::value_type request{ handshake_req(slot_info_) };
 
                 // Make unary request
                 handshake_response::value_type response;
                 status = co_await RPC::request(grpc_context, handshake_stub, client_context, request, response, boost::asio::use_awaitable);
                 handshake_response handshake_rsp;
                 plugin_info = handshake_rsp(response, client_context.peer());
-                valid_ = validator_type{ *slot_info_, plugin_info };
+                valid_ = validator_type{ slot_info_, plugin_info };
                 if (valid_)
                 {
-                    spdlog::info("Using plugin: '{}-{}' running at [{}] for slot {}", plugin_info.plugin_name, plugin_info.plugin_version, plugin_info.peer, slot_info_->slot_id);
+                    spdlog::info("Using plugin: '{}-{}' running at [{}] for slot {}", plugin_info.plugin_name, plugin_info.plugin_version, plugin_info.peer, slot_info_.slot_id);
                     if (! plugin_info.broadcast_subscriptions.empty())
                     {
                         spdlog::info("Subscribing plugin '{}' to the following broadcasts {}", plugin_info.plugin_name, plugin_info.broadcast_subscriptions);
@@ -122,11 +123,11 @@ public:
 
         if (! status.ok()) // TODO: handle different kind of status codes
         {
-            throw exceptions::RemoteException(*slot_info_, status.error_message());
+            throw exceptions::RemoteException(slot_info_, status.error_message());
         }
         if (! plugin_info.plugin_name.empty() && ! plugin_info.slot_version.empty())
         {
-            plugin_info_->emplace(plugin_info);
+            plugin_info_.emplace(plugin_info);
         }
     };
 
@@ -136,21 +137,21 @@ public:
     {
         if (this != &other)
         {
+            invoke_stub_ = other.invoke_stub_;
+            broadcast_stub_ = other.broadcast_stub_;
             valid_ = other.valid_;
-            invoke_component_ = other.invoke_component_;
-            broadcast_component_ = other.broadcast_component_;
             plugin_info_ = other.plugin_info_;
             slot_info_ = other.slot_info_;
         }
         return *this;
     }
-    constexpr PluginProxy& operator=(PluginProxy&& other)
+    constexpr PluginProxy& operator=(PluginProxy&& other) noexcept
     {
         if (this != &other)
         {
+            invoke_stub_ = std::move(other.invoke_stub_);
+            broadcast_stub_ = std::move(other.broadcast_stub_);
             valid_ = std::move(other.valid_);
-            invoke_component_ = std::move(other.invoke_component_);
-            broadcast_component_ = std::move(other.broadcast_component_);
             plugin_info_ = std::move(other.plugin_info_);
             slot_info_ = std::move(other.slot_info_);
         }
@@ -160,25 +161,125 @@ public:
 
     value_type invoke(auto&&... args)
     {
-        return invoke_component_.invoke(std::forward<decltype(args)>(args)...);
+        agrpc::GrpcContext grpc_context;
+        value_type ret_value{};
+        grpc::Status status;
+
+        boost::asio::co_spawn(
+            grpc_context,
+            [this, &grpc_context, &status, &ret_value, &args...]()
+            {
+                return this->invokeCall(grpc_context, status, ret_value, std::forward<decltype(args)>(args)...);
+            },
+            boost::asio::detached);
+        grpc_context.run();
+
+        if (! status.ok()) // TODO: handle different kind of status codes
+        {
+            if (plugin_info_.has_value())
+            {
+                throw exceptions::RemoteException(slot_info_, plugin_info_.value(), status.error_message());
+            }
+            throw exceptions::RemoteException(slot_info_, status.error_message());
+        }
+        return ret_value;
     }
 
     template<plugins::v0::SlotID Subscription>
     void broadcast(auto&&... args)
     {
-        return broadcast_component_.template broadcast<Subscription>(std::forward<decltype(args)>(args)...);
+        if (! plugin_info_.value().broadcast_subscriptions.contains(Subscription))
+        {
+            return;
+        }
+        agrpc::GrpcContext grpc_context;
+        grpc::Status status;
+
+        boost::asio::co_spawn(
+            grpc_context,
+            [this, &grpc_context, &status, &args...]()
+            {
+                return this->broadcastCall<Subscription>(grpc_context, status, std::forward<decltype(args)>(args)...);
+            },
+            boost::asio::detached);
+        grpc_context.run();
+
+        if (! status.ok()) // TODO: handle different kind of status codes
+        {
+            if (plugin_info_.has_value())
+            {
+                throw exceptions::RemoteException(slot_info_, plugin_info_.value(), status.error_message());
+            }
+            throw exceptions::RemoteException(slot_info_, status.error_message());
+        }
     }
 
 private:
+    inline static void prep_client_context(grpc::ClientContext& client_context, const slot_metadata& slot_info, const std::chrono::milliseconds& timeout = std::chrono::minutes(5))
+    {
+        // Set time-out
+        client_context.set_deadline(std::chrono::system_clock::now() + timeout);
+
+        // Metadata
+        client_context.AddMetadata("cura-engine-uuid", slot_info.engine_uuid.data());
+        client_context.AddMetadata("cura-thread-id", fmt::format("{}", std::this_thread::get_id()));
+    }
+
+    /**
+     * @brief Executes the invokeCall operation with the plugin.
+     *
+     * Sends a request to the plugin and saves the response.
+     *
+     * @param grpc_context - The gRPC context to use for the call
+     * @param status - Status of the gRPC call which gets updated in this method
+     * @param ret_value - Reference to the value in which response to be stored
+     * @param args - Request arguments
+     * @return A boost::asio::awaitable<void> indicating completion of the operation
+     */
+    boost::asio::awaitable<void> invokeCall(agrpc::GrpcContext& grpc_context, grpc::Status& status, value_type& ret_value, auto&&... args)
+    {
+        using RPC = agrpc::ClientRPC<&invoke_stub_t::PrepareAsyncCall>;
+        grpc::ClientContext client_context{};
+        prep_client_context(client_context, slot_info_);
+
+        // Construct request
+        auto request{ req_(std::forward<decltype(args)>(args)...) };
+
+        // Make unary request
+        rsp_msg_type response;
+        status = co_await RPC::request(grpc_context, invoke_stub_, client_context, request, response, boost::asio::use_awaitable);
+        ret_value = rsp_(response);
+        co_return;
+    }
+
+    template<plugins::v0::SlotID Subscription>
+    boost::asio::awaitable<void> broadcastCall(agrpc::GrpcContext& grpc_context, grpc::Status& status, auto&&... args)
+    {
+        grpc::ClientContext client_context{};
+        prep_client_context(client_context, slot_info_);
+        using RPC = agrpc::ClientRPC<&broadcast_stub_t::PrepareAsyncBroadcastSettings>;
+
+        details::broadcast_rpc<Subscription, broadcast_stub_t> requester{};
+        auto request = requester(std::forward<decltype(args)>(args)...);
+
+        auto response = google::protobuf::Empty{};
+        status = co_await RPC::request(
+            grpc_context,
+            broadcast_stub_,
+            client_context,
+            request,
+            response,
+            boost::asio::use_awaitable);
+        co_return;
+    }
+
     validator_type valid_{}; ///< The validator object for plugin validation.
-
-    details::slot_info_ptr slot_info_ ///< Holds information about the plugin slot.
-        { std::make_shared<slot_metadata>(slot_metadata{ .slot_id = SlotID, .version_range = SlotVersionRng.value, .engine_uuid = Application::getInstance().instance_uuid }) };
-    details::plugin_info_ptr plugin_info_{ std::make_shared<std::optional<plugin_metadata>>(
-        std::nullopt) }; ///< Optional object that holds the plugin metadata, set after handshake
-
-    invoke_component_t invoke_component_;
-    broadcast_component_t broadcast_component_;
+    req_converter_type req_{}; ///< The Invoke request converter object.
+    rsp_converter_type rsp_{}; ///< The Invoke response converter object.
+    slot_metadata slot_info_{ .slot_id = SlotID,
+                              .version_range = SlotVersionRng.value,
+                              .engine_uuid = Application::getInstance().instance_uuid }; ///< Holds information about the plugin slot.
+    std::optional<plugin_metadata> plugin_info_{ std::optional<plugin_metadata>(std::nullopt) }; ///< Optional object that holds the plugin metadata, set after handshake
 };
 
 } // namespace cura::plugins
