+++ include/plugins/slots.h
@@ -11,6 +11,7 @@
 #include "cura/plugins/slots/postprocess/v0/modify.grpc.pb.h"
 #include "cura/plugins/slots/simplify/v0/modify.grpc.pb.h"
 #include "cura/plugins/v0/slot_id.pb.h"
+#include "infill.h"
 #include "plugins/converters.h"
 #include "plugins/slotproxy.h"
 #include "plugins/types.h"
@@ -22,6 +23,7 @@
 
 #include <exception>
 #include <memory>
+#include <tuple>
 #include <utility>
 
 namespace cura
@@ -49,7 +51,7 @@ struct simplify_default
 
 struct infill_generate_default
 {
-    std::tuple<std::vector<VariableWidthLines>, Polygons, Polygons> operator()(Polygons _infill, auto&&... args)
+    std::tuple<std::vector<VariableWidthLines>, Polygons, Polygons> operator()(auto&&... args)
     {
         // this code is only reachable when no slot is registered while the infill type is requested to be
         // generated by a plugin; this should not be possible to set up in the first place. Return an empty
